+++ tests/test_attributes.py
@@ -1,19 +1,46 @@
+import pytest
+
 from magento import attributes
-from magento.types import Product
-
-custom_attributes1 = [{"attribute_code": "foo", "value": "bar"}]
-custom_attributes2 = [{"attribute_code": "foo", "value": "bar"},
-                      {"attribute_code": "bar", "value": "foo foo"}]
-custom_attributes3 = [{"attribute_code": "yes", "value": "1"},
-                      {"attribute_code": "nope", "value": "0"}]
-
-product0: Product = {"custom_attributes": []}
-product1: Product = {"custom_attributes": custom_attributes1}
-product2: Product = {"custom_attributes": custom_attributes2}
-product3: Product = {"custom_attributes": custom_attributes3}
 
 
-def test_get_custom_attribute():
+@pytest.fixture
+def custom_attributes1():
+    return [{"attribute_code": "foo", "value": "bar"}]
+
+
+@pytest.fixture
+def custom_attributes2():
+    return [{"attribute_code": "foo", "value": "bar"},
+            {"attribute_code": "bar", "value": "foo foo"}]
+
+
+@pytest.fixture
+def custom_attributes3():
+    return [{"attribute_code": "yes", "value": "1"},
+            {"attribute_code": "nope", "value": "0"}]
+
+
+@pytest.fixture
+def product0():
+    return {"custom_attributes": []}
+
+
+@pytest.fixture
+def product1(custom_attributes1):
+    return {"custom_attributes": custom_attributes1}
+
+
+@pytest.fixture
+def product2(custom_attributes2):
+    return {"custom_attributes": custom_attributes2}
+
+
+@pytest.fixture
+def product3(custom_attributes3):
+    return {"custom_attributes": custom_attributes3}
+
+
+def test_get_custom_attribute(product0, product1, product2):
     assert attributes.get_custom_attribute(product0, "foo") is None
     assert attributes.get_custom_attribute(product1, "idontexist") is None
     assert attributes.get_custom_attribute(product2, "idontexist") is None
@@ -47,19 +74,39 @@ def test_get_custom_attribute():
     assert not attributes.get_custom_attribute(product, "false", bool)
 
 
-def test_get_boolean_custom_attribute():
+def test_get_boolean_custom_attribute(product0, product1, product2, product3):
     assert attributes.get_boolean_custom_attribute(product0, "foo") is None
     assert attributes.get_boolean_custom_attribute(product3, "yes") is True
     assert attributes.get_boolean_custom_attribute(product3, "nope") is False
 
 
-def test_get_custom_attributes_dict():
+def test_get_custom_attributes_dict(product0, product1, product2):
     assert attributes.get_custom_attributes_dict(product0) == {}
     assert attributes.get_custom_attributes_dict(product1) == {"foo": "bar"}
     assert attributes.get_custom_attributes_dict(product2) == {"foo": "bar", "bar": "foo foo"}
 
 
-def test_pretty_custom_attributes():
+def test_pretty_custom_attributes(custom_attributes1, custom_attributes2):
     assert attributes.pretty_custom_attributes([]) == ""
     assert attributes.pretty_custom_attributes(custom_attributes1) == "foo='bar'"
     assert attributes.pretty_custom_attributes(custom_attributes2) == "foo='bar', bar='foo foo'"
+
+
+def test_set_custom_attribute_empty_item():
+    assert attributes.set_custom_attribute({}, "a", True) == \
+           {"custom_attributes": [{"attribute_code": "a", "value": "1"}]}
+    assert attributes.set_custom_attribute({}, "a", False) == \
+           {"custom_attributes": [{"attribute_code": "a", "value": "0"}]}
+    assert attributes.set_custom_attribute({}, "a", None) == \
+           {"custom_attributes": [{"attribute_code": "a", "value": ""}]}
+    assert attributes.set_custom_attribute({}, "a", 42) == \
+           {"custom_attributes": [{"attribute_code": "a", "value": "42"}]}
+    assert attributes.set_custom_attribute({}, "a", "b") == \
+           {"custom_attributes": [{"attribute_code": "a", "value": "b"}]}
+
+
+def test_set_custom_attribute(product3):
+    product4 = {}
+    for attribute in product3["custom_attributes"]:
+        product4 = attributes.set_custom_attribute(product4, attribute["attribute_code"], attribute["value"])
+    assert product4 == product3