+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 * Deprecate `get_product_source_items(sku)` in favor of `get_source_items(sku=sku)`, which returns exactly the same data
-* Add `get_stock_source_links`, `get_product_stock_item`, `get_product_stock_status`
+* Add `get_stock_source_links`, `get_product_stock_item` and `get_product_stock_status` methods
+* Add `set_custom_attribute` function
 * Document parameters of `unlink_child_product`
 
 ## 1.3.2 (2022/03/14)